requirements.txt
ffmpeg==1.4
ffmpeg-python==0.2.0
numpy==1.23.4
opencv-python==4.6.0.66
simpleaudio==1.0.4
Pillow==9.2.0
pycryptodomex==3.15.0